requirements.txt
Django==4.2
djangorestframework
gunicorn
django-cors-headers